+++ packages/zent/src/design/Design.js
@@ -48,7 +48,8 @@ import {
   isExpectedDesginType,
   serializeDesignType
 } from './utils/design-type';
-import InstanceCountMap from './utils/InstanceCountMap';
+import LazyMap from './utils/LazyMap';
+import { ADD_COMPONENT_OVERLAY_POSITION } from './constants';
 
 const UUID_KEY = '__zent-design-uuid__';
 const CACHE_KEY = '__zent-design-cache-storage__';
@@ -92,10 +93,16 @@ export default class Design extends (PureComponent || Component) {
         // 组件是否出现在添加组件的列表里面
         appendable: PropTypes.bool,
 
-        // 是否显示右下角的编辑区域(编辑/加内容/删除)
-        // 不支持在这里配置编辑区域的按钮，参数太多。
+        // 是否显示右下角的编辑区域(加内容/删除)
+        // 如果要单独控制删除/加内容，请使用 canDelete 和 canInsert 来控制
         // 如果要自定义编辑区域，可以通过重写 previewController 的方式来做。
         configurable: PropTypes.bool,
+
+        // 是否可以删除
+        canDelete: PropTypes.bool,
+
+        // hover 的时候时候显示添加组件的按钮
+        canInsert: PropTypes.bool,
 
         // 组件是否可以编辑
         // 可以选中的组件一定是可以编辑的
@@ -108,6 +115,9 @@ export default class Design extends (PureComponent || Component) {
 
         // 组件可以添加的最大次数
         limit: PropTypes.oneOfType([PropTypes.number, PropTypes.func]),
+
+        // 组件达到最大添加次数后，鼠标移上去的提示
+        limitMessage: PropTypes.oneOfType([PropTypes.node, PropTypes.func]),
 
         // 是否可以添加组件的回调函数，返回一个 Promise
         shouldCreate: PropTypes.func
@@ -201,6 +211,9 @@ export default class Design extends (PureComponent || Component) {
 
       // 是否显示添加组件的浮层
       showAddComponentOverlay: false,
+
+      // 添加组件浮层的位置
+      addComponentOverlayPosition: ADD_COMPONENT_OVERLAY_POSITION.UNKNOWN,
 
       // 可添加的组件列表
       appendableComponents: [],
@@ -323,6 +336,7 @@ export default class Design extends (PureComponent || Component) {
       selectedUUID,
       appendableComponents,
       showAddComponentOverlay,
+      addComponentOverlayPosition,
       validations,
       showError,
       componentInstanceCount
@@ -341,6 +355,7 @@ export default class Design extends (PureComponent || Component) {
       selectedUUID,
       getUUIDFromValue: this.getUUIDFromValue,
       showAddComponentOverlay,
+      addComponentOverlayPosition,
       onAdd: this.onShowAddComponentOverlay,
       onEdit: this.onShowEditComponentOverlay,
       onSelect: this.onSelect,
@@ -385,12 +400,12 @@ export default class Design extends (PureComponent || Component) {
   };
 
   // 打开右侧添加新组件的弹层
-  onShowAddComponentOverlay = component => {
-    this.toggleEditOrAdd(component, true);
+  onShowAddComponentOverlay = (component, addPosition) => {
+    this.toggleEditOrAdd(component, true, addPosition);
 
     // 将当前组件滚动到顶部
-    const id = this.getUUIDFromValue(component);
-    this.scrollToPreviewItem(id);
+    // const id = this.getUUIDFromValue(component);
+    // this.scrollToPreviewItem(id);
   };
 
   // 编辑一个已有组件
@@ -405,8 +420,9 @@ export default class Design extends (PureComponent || Component) {
   // 选中一个组件
   onSelect = component => {
     const id = this.getUUIDFromValue(component);
+    const { showAddComponentOverlay } = this.state;
 
-    if (this.isSelected(component)) {
+    if (this.isSelected(component) && !showAddComponentOverlay) {
       return;
     }
 
@@ -434,9 +450,16 @@ export default class Design extends (PureComponent || Component) {
     let newValue;
     if (fromSelected) {
       newValue = value.slice();
+      const { addComponentOverlayPosition } = this.state;
       const { selectedUUID } = this.state;
       const selectedIndex = findIndex(value, { [UUID_KEY]: selectedUUID });
-      newValue.splice(selectedIndex + 1, 0, instance);
+
+      // 两种位置，插入到当前选中的前面或者后面
+      const delta =
+        addComponentOverlayPosition === ADD_COMPONENT_OVERLAY_POSITION.TOP
+          ? 0
+          : 1;
+      newValue.splice(selectedIndex + delta, 0, instance);
     } else {
       newValue = value.concat(instance);
     }
@@ -481,14 +504,74 @@ export default class Design extends (PureComponent || Component) {
     });
   };
 
-  // 交换两个组件的位置
   onMove = (fromIndex, toIndex) => {
-    const { value } = this.props;
-    const newValue = value.slice();
+    if (fromIndex === toIndex) {
+      return;
+    }
 
-    const tmp = value[fromIndex];
-    newValue[fromIndex] = newValue[toIndex];
-    newValue[toIndex] = tmp;
+    const { value, components } = this.props;
+    const newValue = [];
+    let tmp;
+
+    /**
+     * 这个算法不是仅仅交换两个位置的节点，所有中间节点都需要移位
+     * 需要考虑数组中间有不可拖拽节点的情况，这种情况下 fromIndex, toIndex 的值是不包括这些节点的
+     * 例如 [1, 0, 0, 1, 0, 0, 1]: fromIndex = 0, toIndex = 1 表示移动第一个和第二个 1。
+     */
+    let passedFromIndex = false;
+    let passedToIndex = false;
+
+    if (fromIndex < toIndex) {
+      for (let i = 0, dragableIndex = -1; i < value.length; i++) {
+        const val = value[i];
+
+        const comp = find(components, c => isExpectedDesginType(c, val.type));
+        const dragable = comp && defaultTo(comp.dragable, true);
+        if (dragable) {
+          dragableIndex++;
+        }
+
+        /* Invariant: Each step copies one value, except one copies 2 and another doesn't copy */
+        if (dragableIndex === fromIndex && !passedFromIndex) {
+          tmp = val;
+          passedFromIndex = true;
+        } else if (dragableIndex < toIndex && passedFromIndex) {
+          newValue[i - 1] = val;
+        } else if (dragableIndex === toIndex && !passedToIndex) {
+          newValue[i - 1] = val;
+          newValue[i] = tmp;
+          passedToIndex = true;
+        } else {
+          newValue[i] = val;
+        }
+      }
+    } else {
+      let toInsetIndex;
+
+      for (let i = 0, dragableIndex = -1; i < value.length; i++) {
+        const val = value[i];
+
+        const comp = find(components, c => isExpectedDesginType(c, val.type));
+        const dragable = comp && defaultTo(comp.dragable, true);
+        if (dragable) {
+          dragableIndex++;
+        }
+
+        /* Invariant: each step copies one value */
+        if (dragableIndex === toIndex && !passedToIndex) {
+          toInsetIndex = i;
+          newValue[i + 1] = val;
+          passedToIndex = true;
+        } else if (dragableIndex < fromIndex && passedToIndex) {
+          newValue[i + 1] = val;
+        } else if (dragableIndex === fromIndex && !passedFromIndex) {
+          newValue[toInsetIndex] = val;
+          passedFromIndex = true;
+        } else {
+          newValue[i] = val;
+        }
+      }
+    }
 
     this.trackValueChange(newValue);
   };
@@ -585,17 +668,26 @@ export default class Design extends (PureComponent || Component) {
     this.removeCache();
   };
 
-  toggleEditOrAdd(component, showAdd) {
-    const { showAddComponentOverlay } = this.state;
+  toggleEditOrAdd(
+    component,
+    showAdd,
+    addPosition = ADD_COMPONENT_OVERLAY_POSITION.UNKNOWN
+  ) {
+    const { showAddComponentOverlay, addComponentOverlayPosition } = this.state;
     const id = this.getUUIDFromValue(component);
 
-    if (this.isSelected(component) && showAddComponentOverlay === showAdd) {
+    if (
+      this.isSelected(component) &&
+      showAddComponentOverlay === showAdd &&
+      addPosition === addComponentOverlayPosition
+    ) {
       return;
     }
 
     this.setState({
       selectedUUID: id,
-      showAddComponentOverlay: showAdd
+      showAddComponentOverlay: showAdd,
+      addComponentOverlayPosition: addPosition
     });
     this.adjustHeight();
   }
@@ -877,7 +969,7 @@ function findFirstEditableSibling(value, components, startIndex) {
  * @param {Array} components Design 支持的组件列表
  */
 function makeInstanceCountMapFromValue(value, components) {
-  const instanceCountMap = new InstanceCountMap(0);
+  const instanceCountMap = new LazyMap(0);
 
   (value || []).forEach(val => {
     const comp = find(components, c => isExpectedDesginType(c, val.type));